Exported the switch statement to it's own file
